// typechain/FXPool.d.ts
export class FXPool extends BaseContract {
// ...
  functions: {
// ...
      sender: string,
      recipient: string,
      balances: BigNumberish[],
      lastChangeBlock: BigNumberish,
      protocolSwapFeePercentage: BigNumberish,
      userData: BytesLike,
      overrides?: Overrides & { from?: string | Promise<string> }
// ...
  };
// ...
}